requirements/base.txt
# Basics
wagtail==5.2.8
--extra-index-url https://__token__:glpat-ZJJHm8J1fE9t-LiRys8q@git.noripyt.com/api/v4/groups/55/-/packages/pypi/simple
noridjango[compressor]==3.2.6

# Database fields
django-tree==0.5.6
django-autoslug==1.9.8

# User interface
django-grappelli==2.15.6
django-super-inlines==0.2.2
django-tinymce==2.9.0
django-crispy-forms==1.12.0
django-ajax-selects==2.1.0
django-image-cropping==1.5.0

# User authentication
django-allauth==0.50.0

# Search engine
django-haystack==3.0
elasticsearch==1.7.0

# Caching
django-redis==5.0.0

# Data visualization frameworks
djangorestframework==3.12.4
django-tablature==0.2.1

# Small but powerful tools
django-reversion==3.0.9
django-el-pagination==3.3.0
easy-thumbnails==2.7.1
django-viewsets==0.2.0
tqdm==4.66.1
awesome-slugify==1.6.5

# Asynchronous tasks
django-rq==2.4.1
rq==1.8.1

# CSS/JS handling
cssmin==0.2.0

# Data parsing/serialization
beautifulsoup4==4.9.3
lxml==4.6.3
defusedxml==0.7.1
pandas==1.3.5
XlsxWriter==1.4.3

# HTML cleaning
bleach==3.3.0

# Dependencies
# For easy-thumbnails
Pillow==10.4.0
# For djangorestframework (optionals)
Pygments==2.9.0
django-filter==23.5
# For exporting DataFrames to XLSX, Pandas needs Jinja2 (yes…)
Jinja2==3.1.4
